create spark_delta_utils
